Create bricks.py
Consider a set of wooden blocks with two types of blocks: 3"x1" and 4.5"x1". We
want to build a wall blocks, but to make it sturdy, the blocks need to overlap completely and be
staggered, and also not hang off the sides, e.g., it should look like this:
 ________________ 
|_______||_______|
|____||____||____|
And not like this:
 ________________ 
|____||____||____|
|____||____||____|

Or this:
 ___________________
|____||_______||____|
|____||____||____|

There are 2 ways to build a 7.5”x1” wall, and also 2 ways to build a 7.5”x2” (that is, 2 blocks high), 4 for 12”x3”, 7,958 for 27”x5”, and so on.

Write a program to determine how many different combinations of bricks can be used to build a 48”x10” wall.